requirements.txt
# Django Core
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Database
mysqlclient>=2.1.0
psycopg2-binary>=2.9.0
dj-database-url>=2.0.0

# Authentication & JWT
djangorestframework-simplejwt>=5.2.0
PyJWT>=2.7.0

# Google OAuth
django-allauth>=0.54.0
dj-rest-auth>=4.0.0

# Redis & Caching
django-redis>=5.3.0
redis>=4.5.0

# Celery for Background Tasks
Celery>=5.3.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# Environment Variables
python-dotenv>=1.0.0

# Image Handling
Pillow>=10.0.0

# Production Server
gunicorn>=21.0.0

# Validation & Utils
phonenumbers>=8.13.0
python-magic>=0.4.27

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
factory-boy>=3.3.0

# Documentation
drf-spectacular>=0.26.0

# Filtering
django-filter>=23.0
